FLUID-6504: Split up BindModel tests into two files
This involved lifting the test grade for BindModel websockets to shared testing utilities
